Split onCreateNode
