+++ www/src/utils/node/creators.js
@@ -1,0 +1,24 @@
+const slugify = require(`slugify`)
+
+exports.onCreateNode = ({ node, actions }) => {
+  const { createNodeField } = actions
+  let slug
+  if (node.internal.type === `CreatorsYaml`) {
+    // Creator pages
+    const validTypes = {
+      individual: `people`,
+      agency: `agencies`,
+      company: `companies`,
+    }
+
+    if (!validTypes[node.type]) {
+      throw new Error(
+        `Creators must have a type of “individual”, “agency”, or “company”, but invalid type “${node.type}” was provided for ${node.name}.`
+      )
+    }
+    slug = `/creators/${validTypes[node.type]}/${slugify(node.name, {
+      lower: true,
+    })}`
+    createNodeField({ node, name: `slug`, value: slug })
+  }
+}
