+++ www/src/utils/node/docs.js
@@ -1,0 +1,100 @@
+const _ = require(`lodash`)
+const path = require(`path`)
+const url = require(`url`)
+const moment = require(`moment`)
+const langs = require("../../../i18n.json")
+
+// convert a string like `/some/long/path/name-of-docs/` to `name-of-docs`
+const slugToAnchor = slug =>
+  slug
+    .split(`/`) // split on dir separators
+    .filter(item => item !== ``) // remove empty values
+    .pop() // take last item
+
+const docSlugFromPath = parsedFilePath => {
+  if (parsedFilePath.name !== `index` && parsedFilePath.dir !== ``) {
+    return `/${parsedFilePath.dir}/${parsedFilePath.name}/`
+  } else if (parsedFilePath.dir === ``) {
+    return `/${parsedFilePath.name}/`
+  } else {
+    return `/${parsedFilePath.dir}/`
+  }
+}
+
+exports.onCreateNode = ({ node, actions, getNode }) => {
+  const { createNodeField } = actions
+  let slug
+  let locale
+  if (node.internal.type === `File`) {
+    const parsedFilePath = path.parse(node.relativePath)
+    // TODO add locale data for non-MDX files
+    if (node.sourceInstanceName === `docs`) {
+      slug = docSlugFromPath(parsedFilePath)
+    }
+    if (slug) {
+      createNodeField({ node, name: `slug`, value: slug })
+    }
+  } else if (
+    [`MarkdownRemark`, `Mdx`].includes(node.internal.type) &&
+    getNode(node.parent).internal.type === `File`
+  ) {
+    const fileNode = getNode(node.parent)
+    const parsedFilePath = path.parse(fileNode.relativePath)
+    // Add slugs for docs pages
+    if (fileNode.sourceInstanceName === `docs`) {
+      slug = docSlugFromPath(parsedFilePath)
+      locale = "en"
+
+      // Set released status and `published at` for blog posts.
+      if (_.includes(parsedFilePath.dir, `blog`)) {
+        let released = false
+        const date = _.get(node, `frontmatter.date`)
+        if (date) {
+          released = moment.utc().isSameOrAfter(moment.utc(date))
+        }
+        createNodeField({ node, name: `released`, value: released })
+
+        const canonicalLink = _.get(node, `frontmatter.canonicalLink`)
+        const publishedAt = _.get(node, `frontmatter.publishedAt`)
+
+        createNodeField({
+          node,
+          name: `publishedAt`,
+          value: canonicalLink
+            ? publishedAt || url.parse(canonicalLink).hostname
+            : null,
+        })
+      }
+    }
+
+    for (let { code } of langs) {
+      if (fileNode.sourceInstanceName === `docs-${code}`) {
+        // have to remove the beginning "/docs" path because of the way
+        // gatsby-source-filesystem and gatsby-source-git differ
+        slug = docSlugFromPath(path.parse(fileNode.relativePath.substring(5)))
+        locale = code
+      }
+    }
+
+    // Add slugs for package READMEs.
+    if (
+      fileNode.sourceInstanceName === `packages` &&
+      parsedFilePath.name === `README`
+    ) {
+      slug = `/packages/${parsedFilePath.dir}/`
+      createNodeField({
+        node,
+        name: `title`,
+        value: parsedFilePath.dir,
+      })
+      createNodeField({ node, name: `package`, value: true })
+    }
+    if (slug) {
+      createNodeField({ node, name: `anchor`, value: slugToAnchor(slug) })
+      createNodeField({ node, name: `slug`, value: slug })
+    }
+    if (locale) {
+      createNodeField({ node, name: `locale`, value: locale })
+    }
+  }
+}
